댓글 조회까지 완료

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -1,10 +1,12 @@
 package com.example.prj1be.controller;
 
 import com.example.prj1be.domain.Board;
+import com.example.prj1be.domain.Member;
 import com.example.prj1be.domain.MyDto1;
 import com.example.prj1be.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.apache.ibatis.annotations.Delete;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,11 +20,15 @@ public class BoardController {
     private final BoardService service;
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board) {
-        if (!service.validate(board)) {
+    public ResponseEntity add(@RequestBody MyDto1 dto, @SessionAttribute(value = "login",required = false) Member login) {
+        if (login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.validate(dto)) {
             return ResponseEntity.badRequest().build();
         }
-        if (service.save(board)) {
+
+        if (service.save(dto,login)) {
             return ResponseEntity.ok().build();
         } else {
             return ResponseEntity.internalServerError().build();
@@ -41,7 +47,14 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id, @SessionAttribute(value = "login",required = false) Member login) {
+
+        if (login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(id,login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
@@ -50,13 +63,20 @@ public class BoardController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody MyDto1 dto) {
+    public ResponseEntity edit(@RequestBody MyDto1 dto, @SessionAttribute(value = "login",required = false)Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(dto.getId(),login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         if (service.edit(dto)) {
             return ResponseEntity.ok().build();
         } else {
             return ResponseEntity.internalServerError().build();
         }
     }
+
